requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
pillow==10.2.0
pydantic==2.11.3
pydantic_core==2.33.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
